api_launcher/requirements.txt
asyncpg
fastapi
uvicorn[standard]
nats-py
pyzmq
prometheus-client
python-dateutil